requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
sendgrid==6.9.7
Werkzeug==2.3.6
PyJWT==2.7.0
pytest==7.4.2
flask-login==0.6.2
flask-mail==0.9.1
python-dotenv==1.0.0
requests==2.31.0
python-whois==0.8.0
dnspython==2.4.2